ASNBlock: Refactor expiry logic, allow more configuration
Refactor prefilled block expiry logic from make_section into get_expiry
for better testing. The default expiry for a site can be configured in
the site config. The expiry for a provider can also be configured. Both
take a string, which will be used exactly, or a sequence of two integer
numbers of months that will be passed to random.randint.