ASNBlock: Query only first range address in RIPEstat WHOIS lookup
RIPEstat supports range WHOIS lookups, but if the requested range is
larger than the WHOIS range, it will return results for a larger range.
That means that the WHOIS filtering is less specific. Requesting only
the first address will give the result for the first WHOIS range in the
range we're looking at.

In the future, it may be useful to iterate through all the WHOIS results
in the range based on the `inetnum` value, and only include the
sub-ranges matching a result. That would substantially increase the
un-cached WHOIS lookup time.